Moved RemTicketFields back to REM.
